Move import spots from label image plugin to org.mastodon.mamut.io.importer.labelimage
* These respective plugin is importing instead of exporting
